Adjust module load transactions shares from report files to manage usd and mxn currency

+++ src/main/java/project/stock/track/app/beans/pojos/business/transaction/TransactionIssueFilePojo.java
@@ -14,7 +14,7 @@ public class TransactionIssueFilePojo {
 	
 	private String typeTransaction;
 	
-	private double titles;
+	private BigDecimal titles;
 	
 	private BigDecimal price;
